wsd: cleanup: moved all parsing related static methods and containers to HostUtil.cpp
