+++ Makefile.am
@@ -130,6 +130,7 @@ coolwsd_sources = common/Crypto.cpp \
                   wsd/FileServerUtil.cpp \
                   wsd/RequestDetails.cpp \
                   wsd/Storage.cpp \
+                  wsd/HostUtil.cpp \
                   wsd/TileCache.cpp \
                   wsd/ProofKey.cpp \
                   wsd/QuarantineUtil.cpp
@@ -263,7 +264,8 @@ wsd_headers = wsd/Admin.hpp \
               wsd/TileDesc.hpp \
               wsd/TraceFile.hpp \
               wsd/UserMessages.hpp \
-              wsd/QuarantineUtil.hpp
+              wsd/QuarantineUtil.hpp \
+              wsd/HostUtil.hpp
 
 shared_headers = common/Common.hpp \
                  common/Clipboard.hpp \
