+++ wsd/Storage.cpp
@@ -53,6 +53,7 @@
 #include <common/TraceEvent.hpp>
 #include <NetUtil.hpp>
 #include <CommandControl.hpp>
+#include "HostUtil.hpp"
 
 #ifdef IOS
 #include <ios.h>
@@ -60,16 +61,9 @@
 #include "androidapp.hpp"
 #endif
 
-#if !defined(BUILDING_TESTS)
 bool StorageBase::FilesystemEnabled;
-bool StorageBase::WopiEnabled;
 bool StorageBase::SSLAsScheme = true;
 bool StorageBase::SSLEnabled = false;
-Util::RegexListMatcher StorageBase::WopiHosts;
-std::map<std::string, std::string> StorageBase::AliasHosts;
-std::set<std::string> StorageBase::AllHosts;
-std::string StorageBase::FirstHost;
-#endif // !defined(BUILDING_TESTS)
 
 #if !MOBILEAPP
 
@@ -88,145 +82,7 @@ std::string StorageBase::getLocalRootPath() const
 
     return rootPath.toString();
 }
-
-#if !defined(BUILDING_TESTS)
-void StorageBase::parseWopiHost(Poco::Util::LayeredConfiguration& conf)
-{
-    // Parse the WOPI settings.
-    WopiHosts.clear();
-    WopiEnabled = conf.getBool("storage.wopi[@allow]", false);
-    if (WopiEnabled)
-    {
-        for (size_t i = 0;; ++i)
-        {
-            const std::string path = "storage.wopi.host[" + std::to_string(i) + ']';
-            if (!conf.has(path))
-            {
-                break;
-            }
-            StorageBase::addWopiHost(conf.getString(path, ""), conf.getBool(path + "[@allow]", false));
-        }
-    }
-}
-
-void StorageBase::addWopiHost(std::string host, bool allow)
-{
-    if (!host.empty())
-    {
-        if (allow)
-        {
-            LOG_INF("Adding trusted WOPI host: [" << host << "].");
-            WopiHosts.allow(host);
-        }
-        else
-        {
-            LOG_INF("Adding blocked WOPI host: [" << host << "].");
-            WopiHosts.deny(host);
-        }
-    }
-}
-
-void StorageBase::parseAliases(Poco::Util::LayeredConfiguration& conf)
-{
-    //set alias_groups mode to compat
-    if (!conf.has("storage.wopi.alias_groups"))
-    {
-        conf.setString("storage.wopi.alias_groups[@mode]", "compat");
-    }
-    else if (conf.has("storage.wopi.alias_groups.group[0]"))
-    {
-        // group defined in alias_groups
-        if (Util::iequal(config::getString("storage.wopi.alias_groups[@mode]", "first"), "first"))
-        {
-            LOG_ERR("Admins didnot set the alias_groups mode to 'groups'");
-            AliasHosts.clear();
-            AllHosts.clear();
-            return;
-        }
-    }
-
-    AliasHosts.clear();
-    AllHosts.clear();
-
-    for (size_t i = 0;; i++)
-    {
-        const std::string path = "storage.wopi.alias_groups.group[" + std::to_string(i) + ']';
-        if (!conf.has(path + ".host"))
-        {
-            break;
-        }
-
-        const std::string uri = conf.getString(path + ".host", "");
-        if (uri.empty())
-        {
-            continue;
-        }
-        bool allow = conf.getBool(path + ".host[@allow]", false);
-
-        try
-        {
-            const Poco::URI realUri(uri);
-            StorageBase::addWopiHost(realUri.getHost(), allow);
-            AllHosts.insert(realUri.getAuthority());
-        }
-        catch (const Poco::Exception& exc)
-        {
-            LOG_WRN("parseAliases: " << exc.displayText());
-        }
-
-        for (size_t j = 0;; j++)
-        {
-            const std::string aliasPath = path + ".alias[" + std::to_string(j) + ']';
-            if (!conf.has(aliasPath))
-            {
-                break;
-            }
-
-            try
-            {
-                const Poco::URI aliasUri(conf.getString(aliasPath, ""));
-                if (aliasUri.empty())
-                {
-                    continue;
-                }
-                const Poco::URI realUri(uri);
-                AliasHosts.insert({ aliasUri.getAuthority(), realUri.getAuthority() });
-                AllHosts.insert(aliasUri.getAuthority());
-                StorageBase::addWopiHost(aliasUri.getHost(), allow);
-            }
-            catch (const Poco::Exception& exc)
-            {
-                LOG_WRN("parseAliases: " << exc.displayText());
-            }
-        }
-    }
-}
-
-std::string StorageBase::getNewUri(const Poco::URI& uri)
-{
-    if (Util::iequal(config::getString("storage.wopi.alias_groups[@mode]", "first"), "compat"))
-    {
-        return uri.getPath();
-    }
-    Poco::URI newUri(uri);
-    const std::string key = newUri.getAuthority();
-    if (Util::matchRegex(AliasHosts, key))
-    {
-        newUri.setAuthority(AliasHosts[key]);
-    }
-
-    if (newUri.getAuthority().empty())
-    {
-        return newUri.getPath();
-    }
-    return newUri.getScheme() + "://" + newUri.getHost() + ':' + std::to_string(newUri.getPort()) +
-           newUri.getPath();
-}
-#endif // !defined(BUILDING_TESTS)
-
 #endif
-
-#if !defined(BUILDING_TESTS)
 
 void StorageBase::initialize()
 {
@@ -234,13 +90,13 @@ void StorageBase::initialize()
     const auto& app = Poco::Util::Application::instance();
     FilesystemEnabled = app.config().getBool("storage.filesystem[@allow]", false);
 
-    parseWopiHost(app.config());
-
-    parseAliases(app.config());
+    HostUtil::parseWopiHost(app.config());
 
 #ifdef ENABLE_FEATURE_LOCK
     CommandControl::LockManager::parseLockedHost(app.config());
 #endif
+
+    HostUtil::parseAliases(app.config());
 
 #if ENABLE_SSL
     // FIXME: should use our own SSL socket implementation here.
@@ -305,39 +161,6 @@ void StorageBase::initialize()
 #else
     FilesystemEnabled = true;
 #endif
-}
-
-bool StorageBase::allowedWopiHost(const std::string& host)
-{
-    return WopiEnabled && WopiHosts.match(host);
-}
-
-bool StorageBase::allowedAlias(const Poco::URI& uri)
-{
-    if (Util::iequal(config::getString("storage.wopi.alias_groups[@mode]", "first"), "compat"))
-    {
-        return true;
-    }
-
-    if (AllHosts.empty())
-    {
-        if (FirstHost.empty())
-        {
-            FirstHost = uri.getAuthority();
-        }
-        else if (FirstHost != uri.getAuthority())
-        {
-            LOG_ERR("Only allowed host is: " << FirstHost);
-            return false;
-        }
-    }
-    else if (!Util::matchRegex(AllHosts, uri.getAuthority()))
-    {
-        LOG_ERR("Host: " << uri.getAuthority()
-                         << " is not allowed, It is not part of alias_groups configuration");
-        return false;
-    }
-    return true;
 }
 
 #if !MOBILEAPP
@@ -407,12 +230,12 @@ std::unique_ptr<StorageBase> StorageBase::create(const Poco::URI& uri, const std
         LOG_ERR("Local Storage is disabled by default. Enable in the config file or on the command-line to enable.");
     }
 #if !MOBILEAPP
-    else if (WopiEnabled)
+    else if (HostUtil::isWopiEnabled())
     {
         LOG_INF("Public URI [" << COOLWSD::anonymizeUrl(uri.toString()) << "] considered WOPI.");
         const auto& targetHost = uri.getHost();
         bool allowed(false);
-        if ((StorageBase::allowedWopiHost(targetHost) && StorageBase::allowedAlias(uri)) ||
+        if ((HostUtil::allowedWopiHost(targetHost) && HostUtil::allowedAlias(uri)) ||
             isLocalhost(targetHost))
         {
             allowed = true;
@@ -423,8 +246,8 @@ std::unique_ptr<StorageBase> StorageBase::create(const Poco::URI& uri, const std
             const auto hostAddresses(Poco::Net::DNS::resolve(targetHost));
             for (auto &address : hostAddresses.addresses())
             {
-                if (StorageBase::allowedWopiHost(address.toString()) &&
-                    StorageBase::allowedAlias(uri))
+                if (HostUtil::allowedWopiHost(address.toString()) &&
+                    HostUtil::allowedAlias(uri))
                 {
                     allowed = true;
                     break;
@@ -1051,7 +874,7 @@ WopiStorage::WOPIFileInfo::WOPIFileInfo(const FileInfo &fileInfo,
         }
         else
         {
-            LOG_INF("Could not find matching locked host so applying fallback settings");
+            LOG_INF("Could not find matching locked_host: " << host << ",applying fallback settings");
             isReadOnly = config::getBool("feature_lock.locked_hosts.fallback[@read_only]", false);
             isUserLocked =
                 config::getBool("feature_lock.locked_hosts.fallback[@disabled_commands]", false);
@@ -1691,5 +1514,4 @@ WopiStorage::handleUploadToStorageResponse(const WopiUploadDetails& details,
 
 #endif // !MOBILEAPP
 
-#endif // !defined(BUILDING_TESTS)
 /* vim:set shiftwidth=4 softtabstop=4 expandtab: */
